Move error types to seperate module
