+++ src/dxt.rs
@@ -10,7 +10,8 @@
 use std::io::{self, Read, Seek, SeekFrom, Write};
 
 use color::ColorType;
-use image::{self, ImageDecoder, ImageDecoderExt, ImageError, ImageReadBuffer, ImageResult, Progress};
+use image::{self, ImageDecoder, ImageDecoderExt, ImageReadBuffer, Progress};
+use error::{ImageError, ImageResult};
 
 /// What version of DXT compression are we using?
 /// Note that DXT2 and DXT4 are left away as they're
